# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 6016ab4ba3e049112323e4ce
  title: CMO
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /creditline:
    post:
      x-skaffolder-id: 6016b820a3e049112323e5cf
      x-skaffolder-name: create
      x-skaffolder-id-resource: 6016b771a3e049112323e5a6
      x-skaffolder-resource: CreditLine
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - CreditLine
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLine"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                CreditLimit:
                  type: integer
                EndDate:
                  type: integer
                GroupCreditLimit:
                  type: integer
                StartDate:
                  type: integer
                Type:
                  type: string
    get:
      x-skaffolder-id: 6016b821a3e049112323e5d4
      x-skaffolder-name: list
      x-skaffolder-id-resource: 6016b771a3e049112323e5a6
      x-skaffolder-resource: CreditLine
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - CreditLine
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditLine"
  "/creditline/{id}":
    delete:
      x-skaffolder-id: 6016b821a3e049112323e5d5
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6016b771a3e049112323e5a6
      x-skaffolder-resource: CreditLine
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - CreditLine
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6016b820a3e049112323e5ce
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6016b771a3e049112323e5a6
      x-skaffolder-resource: CreditLine
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - CreditLine
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLine"
    post:
      x-skaffolder-id: 6016b820a3e049112323e5d0
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6016b771a3e049112323e5a6
      x-skaffolder-resource: CreditLine
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - CreditLine
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLine"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditLine"
  "/creditline/findByOrganization/{key}":
    get:
      x-skaffolder-id: 6016b80ca3e049112323e5c8
      x-skaffolder-name: findByOrganization
      x-skaffolder-id-resource: 6016b771a3e049112323e5a6
      x-skaffolder-resource: CreditLine
      x-skaffolder-crudAction: findByOrganization
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByOrganization
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByOrganization/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByOrganization
      tags:
        - CreditLine
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditLine"
...
  /organization:
    post:
      x-skaffolder-id: 6016b80ba3e049112323e5c3
      x-skaffolder-name: create
      x-skaffolder-id-resource: 6016b38ca3e049112323e58c
      x-skaffolder-resource: Organization
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Organization
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Active:
                  type: boolean
                ExternalID:
                  type: string
                FormalName:
                  type: string
                Liable:
                  type: boolean
                SearchName:
                  type: string
                Type:
                  type: string
    get:
      x-skaffolder-id: 6016b80ca3e049112323e5c7
      x-skaffolder-name: list
      x-skaffolder-id-resource: 6016b38ca3e049112323e58c
      x-skaffolder-resource: Organization
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Organization
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
  "/organization/{id}":
    delete:
      x-skaffolder-id: 6016b80ca3e049112323e5cb
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6016b38ca3e049112323e58c
      x-skaffolder-resource: Organization
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Organization
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6016b80ba3e049112323e5c2
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6016b38ca3e049112323e58c
      x-skaffolder-resource: Organization
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Organization
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    post:
      x-skaffolder-id: 6016b80ba3e049112323e5c5
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6016b38ca3e049112323e58c
      x-skaffolder-resource: Organization
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Organization
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
  "/organization/findByLessor/{key}":
    get:
      x-skaffolder-id: 6016b7dba3e049112323e5b2
      x-skaffolder-name: findByLessor
      x-skaffolder-id-resource: 6016b38ca3e049112323e58c
      x-skaffolder-resource: Organization
      x-skaffolder-crudAction: findByLessor
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByLessor
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByLessor/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByLessor
      tags:
        - Organization
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreditLine:
      x-skaffolder-id: 6016b771a3e049112323e5a6
      x-skaffolder-id-db: 6016ab4ba3e049112323e4cf
      x-skaffolder-id-entity: 6016b770a3e049112323e59d
      x-skaffolder-url: /creditline
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        CreditLimit:
          type: integer
          x-skaffolder-id-attr: 6016b770a3e049112323e5a1
          x-skaffolder-type: Integer
          x-skaffolder-required: true
        EndDate:
          type: integer
          x-skaffolder-id-attr: 6016b770a3e049112323e59f
          x-skaffolder-type: Date
          x-skaffolder-required: true
        GroupCreditLimit:
          type: integer
          x-skaffolder-id-attr: 6016b770a3e049112323e5a2
          x-skaffolder-type: Integer
        StartDate:
          type: integer
          x-skaffolder-id-attr: 6016b770a3e049112323e59e
          x-skaffolder-type: Date
          x-skaffolder-required: true
        Type:
          type: string
          x-skaffolder-id-attr: 6016b770a3e049112323e5a0
          x-skaffolder-type: String
          x-skaffolder-required: true
          x-skaffolder-enumeration:
            - NonRevolving
            - Revolving
      x-skaffolder-relations:
        Organization:
          x-skaffolder-id: 6016b770a3e049112323e5a3
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6016b770a3e049112323e59d
          x-skaffolder-ent2: 6016b38ca3e049112323e581
      required:
        - _id
        - CreditLimit
        - EndDate
        - StartDate
        - Type
...
    Organization:
      x-skaffolder-id: 6016b38ca3e049112323e58c
      x-skaffolder-id-db: 6016ab4ba3e049112323e4cf
      x-skaffolder-id-entity: 6016b38ca3e049112323e581
      x-skaffolder-url: /organization
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Active:
          type: boolean
          x-skaffolder-id-attr: 6016b38ca3e049112323e586
          x-skaffolder-type: Boolean
        ExternalID:
          type: string
          x-skaffolder-id-attr: 6016b38ca3e049112323e582
          x-skaffolder-type: String
          x-skaffolder-required: true
          x-skaffolder-unique: true
        FormalName:
          type: string
          x-skaffolder-id-attr: 6016b38ca3e049112323e584
          x-skaffolder-type: String
        Liable:
          type: boolean
          x-skaffolder-id-attr: 6016b407a3e049112323e58e
          x-skaffolder-type: Boolean
        SearchName:
          type: string
          x-skaffolder-id-attr: 6016b38ca3e049112323e583
          x-skaffolder-type: String
        Type:
          type: string
          x-skaffolder-id-attr: 6016b38ca3e049112323e585
          x-skaffolder-type: String
          x-skaffolder-enumeration:
            - Client
            - Prospect
            - ProspectBO
      x-skaffolder-relations:
        Country:
          x-skaffolder-id: 6016b38ca3e049112323e587
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6016b38ca3e049112323e581
          x-skaffolder-ent2: 6016b096a3e049112323e56c
        Lessor:
          x-skaffolder-id: 6016b38ca3e049112323e588
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6016b38ca3e049112323e581
          x-skaffolder-ent2: 6016b117a3e049112323e572
        LiableParent:
          x-skaffolder-id: 6016b407a3e049112323e591
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6016b38ca3e049112323e581
          x-skaffolder-ent2: 6016b38ca3e049112323e581
        Parent:
          x-skaffolder-id: 6016b407a3e049112323e58f
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6016b38ca3e049112323e581
          x-skaffolder-ent2: 6016b38ca3e049112323e581
        UtlimateParent:
          x-skaffolder-id: 6016b407a3e049112323e590
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6016b38ca3e049112323e581
          x-skaffolder-ent2: 6016b38ca3e049112323e581
      required:
        - _id
        - ExternalID
...